requirements.txt
requests
pandas
numpy
python-binance